Copied over code from form1.cs to worltrackform.cs and renamed class accordingly
